pyproject.toml
[project]
name = "wa-tfm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.49.0",
    "fastapi>=0.115.12",
    "fire>=0.7.0",
    "httpx>=0.28.1",
    "mcp[cli]>=1.6.0",
    "openai>=1.70.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "supabase>=2.15.0",
]
